move zod validations to separate directory
